Impute missing values in a dataframe and add missingness indicators.

Impute missing values, using knn by default or alternatively median-impute numerics, mode-impute factors. Add missingness indicators.

impute_missing_values(data, type = "standard", add_indicators = T,
  prefix = "miss_", skip_vars = NULL, all_vars = F, remove_constant = T,
  remove_collinear = T, verbose = F)

Arguments

data
Dataframe or matrix.

type
"knn" or "standard" (median/mode). NOTE: knn will result in the data being centered and scaled!

add_indicators
Add a series of missingness indicators.

prefix
String to add at the beginning of the name of each missingness indicator.

skip_vars
List of variable names to exclude from the imputation.

all_vars
Calculate imputation value for all variables, in cases where the imputation info may be used for future datasets.

remove_constant
Remove constant missingness indicators, if applicable.

remove_collinear
Remove collinear missingness indicators, if applicable.

verbose
If True display extra information during execution.

Value

List with the following elements:

$data - imputed dataset.

$impute_info - if knn, caret preprocess element for imputing test data.

$impute_values - if standard, list of imputation values for each variable.

See also

missingness_indicators preProcess

Examples

# Load a test dataset.
data(PimaIndiansDiabetes2, package = "mlbench")

# Check for missing values.
colSums(is.na(PimaIndiansDiabetes2))

#> pregnant  glucose pressure  triceps  insulin     mass pedigree      age 
#>        0        5       35      227      374       11        0        0 
#> diabetes 
#>        0 

# Impute missing data and add missingness indicators.
# Don't impute the outcome though.
result = impute_missing_values(PimaIndiansDiabetes2, skip_vars = "diabetes")

# Confirm we have no missing data.
colSums(is.na(result$data))

#>      pregnant       glucose      pressure       triceps       insulin 
#>             0             0             0             0             0 
#>          mass      pedigree           age      diabetes  miss_glucose 
#>             0             0             0             0             0 
#> miss_pressure  miss_triceps  miss_insulin     miss_mass 
#>             0             0             0             0 

#############
# K-nearest neighbors imputation

result2 = impute_missing_values(PimaIndiansDiabetes2, type = "knn",
                                skip_vars = "diabetes")

# Confirm we have no missing data.
colSums(is.na(result2$data))

#>      pregnant       glucose      pressure       triceps       insulin 
#>             0             0             0             0             0 
#>          mass      pedigree           age      diabetes  miss_glucose 
#>             0             0             0             0             0 
#> miss_pressure  miss_triceps  miss_insulin     miss_mass 
#>             0             0             0             0