Return matrix of missingness indicators for a dataframe or matrix.

Return matrix of missingness indicators for a dataframe or matrix. Removes constant or collinear indicators.

missingness_indicators(data, prefix = "miss_", remove_constant = T,
  remove_collinear = T, skip_vars = c(), verbose = F)

Arguments

data
Dataframe or matrix to analyze for missingness.

prefix
Name prefix for new indicator columns.

remove_constant
Remove any indicators that are all 0 or all 1.

remove_collinear
Remove any indicators that are collinear with each other.

skip_vars
Vector of variable names to skip.

verbose
If TRUE, print additional information.

Value

Matrix of missingness indicators

See also

impute_missing_values

Examples

# Load a test dataset.
data(PimaIndiansDiabetes2, package = "mlbench")

# Check for missing values.
colSums(is.na(PimaIndiansDiabetes2))

#> pregnant  glucose pressure  triceps  insulin     mass pedigree      age 
#>        0        5       35      227      374       11        0        0 
#> diabetes 
#>        0 

# Generate missingness indicators; skip outcome variable.
indicators = missingness_indicators(PimaIndiansDiabetes2, skip_vars = "diabetes")

# Check missingness.
colSums(indicators)

#>  miss_glucose miss_pressure  miss_triceps  miss_insulin     miss_mass 
#>             5            35           227           374            11