Count the terminal nodes in each tree from a random forest

Returns a vector of terminal node counts for each tree in a random forest. The distribution of terminal node counts is helpful when seeking to optimize the maxnodes hyperparameter of the random forest. By default RF allows very large trees, which may result in overfitting. Optimizing the number of terminal nodes in a random forest is a more direct way of requiring simpler trees than the minimum node size hyperparameter.

rf_count_terminal_nodes(rf)

Arguments

rf
Random Forest object

Value

vector of terminal node counts

References

Breiman, L. (2001). Random forests. Machine learning, 45(1), 5-32.

See also

getTree randomForest

Examples

library(SuperLearner)
library(ck37r)

data(Boston, package = "MASS")

set.seed(1)

# Downsample to 100 observations speed up example.
Boston = Boston[sample(nrow(Boston), 100L), ]

sl = SuperLearner(Boston$medv, subset(Boston, select = -medv),
                  family = gaussian(),
                  cvControl = list(V = 3),
                  SL.library = c("SL.mean", "SL.glm", "SL.randomForest"))

#> Loading required package: randomForest

#> randomForest 4.6-12

#> Type rfNews() to see new features/changes/bug fixes.

sl

#> 
#> Call:  
#> SuperLearner(Y = Boston$medv, X = subset(Boston, select = -medv), family = gaussian(),  
#>     SL.library = c("SL.mean", "SL.glm", "SL.randomForest"), cvControl = list(V = 3)) 
#> 
#> 
#> 
#>                         Risk Coef
#> SL.mean_All         82.41584    0
#> SL.glm_All          55.88575    0
#> SL.randomForest_All 34.87371    1

summary(rf_count_terminal_nodes(sl$fitLibrary$SL.randomForest_All$object))

#>    Min. 1st Qu.  Median    Mean 3rd Qu.    Max. 
#>    26.0    32.0    33.0    33.3    35.0    41.0 

max_terminal_nodes =
       max(rf_count_terminal_nodes(sl$fitLibrary$SL.randomForest_All$object))

max_terminal_nodes

#> [1] 41

# Now run create.Learner() based on that maximum.

# It is often handy to convert to log scale of a hyperparameter before
# testing a ~linear grid.
# NOTE: -0.7 ~ 0.69 ~ log(0.5) which is the multiplier that yields sqrt(max)
maxnode_seq = unique(round(exp(log(max_terminal_nodes) *
                               exp(c(-0.97, -0.7, -0.45, -0.15, 0)))))
maxnode_seq

#> [1]  4  6 11 24 41

rf = SuperLearner::create.Learner("SL.randomForest", detailed_names = TRUE,
                                  name_prefix = "rf",
                                  # fewer trees for testing speed only.
                                  params = list(ntree = 100),
                                  tune = list(maxnodes = maxnode_seq))

sl = SuperLearner(Boston$medv, subset(Boston, select = -medv),
                  family = gaussian(),
                  cvControl = list(V = 3),
                  SL.library = c("SL.mean", "SL.glm", rf$names))

sl

#> 
#> Call:  
#> SuperLearner(Y = Boston$medv, X = subset(Boston, select = -medv), family = gaussian(),  
#>     SL.library = c("SL.mean", "SL.glm", rf$names), cvControl = list(V = 3)) 
#> 
#> 
#>                 Risk      Coef
#> SL.mean_All 82.20832 0.0000000
#> SL.glm_All  44.96435 0.0000000
#> rf_4_All    32.67405 0.0000000
#> rf_6_All    26.56812 0.0000000
#> rf_11_All   23.40124 0.2861234
#> rf_24_All   23.07783 0.3243657
#> rf_41_All   23.10118 0.3895109