Setup TMLE to run in parallel

Starts a cluster and sets up TMLE and SuperLearner to use the cluster so that TMLE is conducted in parallel rather than using only one core.

setup_parallel_tmle(parallel = "multicore", max_cores = NULL,
  allow_multinode = T, env = .GlobalEnv)

Arguments

parallel
"multicore", "doParallel", or "doSNOW"

max_cores
Restrict how many many cores will be used on a machine, rather than using all available cores. Useful if each core needs to use a substantial amount of memory.

allow_multinode
If T, will create a multinode cluster if it finds multiple machines listed in the "SLURM_NODELIST" environmental variable. If F, it will only use the current node even if multiple nodes are detected.

env
Environment in which to save the functions, defaulting to the global environment. Set to NULL to disable.

See also

parallelize, tmle_parallel, gen_superlearner